VPP: Added event for VPP connection lost and new/del link.
Now, the PPPoE service starts/stops on the event of a new/del
link using netlink. Added event on VPP connection lost -
VPP-related PPPoE services are stopped. Added "list of monitored interfaces" -
the list on which the PPPoE service would be started on a new link event.
General PPPoE commands to add/del service on an interface, also adds
or dels from "monitored list". To add an interface to the "monitored list",
the interface should be in the config file or added with the command
"pppoe interface add". To remove an interface from the "monitored list" -
the command "pppoe interface del" is used.